build

+++ src/main/java/com/iu/main/employees/EmployeeDTO.java
@@ -3,15 +3,22 @@ package com.iu.main.employees;
 import java.sql.Date;
 
 public class EmployeeDTO {
+	//int와 같은 primitive보단 Integer인 reference로
+	//자바는 보통 카멜케이스 표기법 으로 쓰지만
+	//db의 컬럼명과 동일하게 하기 위해 스네이크 표기법 사용
 	private Integer employee_id;
 	private String first_name;
 	private String last_name;
 	private String eamil;
 	private String phone_number;
+	//자바 Date 
+	//1)java.util 2)java.sql
+	//1)은 deprecated되어 있으므로 2)로 사용
+	//2)는 1)을 상속받음
 	private Date hire_date;
 	private String job_id;
-	private Integer	salary;
-	private Integer commision_pct;
+	private double salary;
+	private double commision_pct;
 	private Integer manager_id;
 	private Integer department_id;
 	public Integer getEmployee_id() {
@@ -56,16 +63,16 @@ public class EmployeeDTO {
 	public void setJob_id(String job_id) {
 		this.job_id = job_id;
 	}
-	public Integer getSalary() {
+	public double getSalary() {
 		return salary;
 	}
-	public void setSalary(Integer salary) {
+	public void setSalary(double salary) {
 		this.salary = salary;
 	}
-	public Integer getCommision_pct() {
+	public double getCommision_pct() {
 		return commision_pct;
 	}
-	public void setCommision_pct(Integer commision_pct) {
+	public void setCommision_pct(double commision_pct) {
 		this.commision_pct = commision_pct;
 	}
 	public Integer getManager_id() {
